add stats for rank options question
