+++ src/web/types/question-details-impl/feedback-rank-options-question-details.impl.ts
@@ -1,3 +1,5 @@
+// tslint:disable-next-line:max-line-length
+import { RankOptionsQuestionStatisticsCalculation } from '../../app/components/question-types/question-statistics/question-statistics-calculation/rank-options-question-statistics-calculation';
 import {
   FeedbackQuestionType,
   FeedbackRankOptionsQuestionDetails, QuestionOutput,
@@ -32,9 +34,29 @@ export class FeedbackRankOptionsQuestionDetailsImpl extends AbstractFeedbackQues
     return ['Feedback', ...optionsHeader];
   }
 
-  getQuestionCsvStats(_: QuestionOutput): string[][] {
-    // TODO
-    return [];
+  getQuestionCsvStats(question: QuestionOutput): string[][] {
+    const statsRows: string[][] = [];
+
+    const statsCalculation: RankOptionsQuestionStatisticsCalculation
+        = new RankOptionsQuestionStatisticsCalculation(this);
+    this.populateQuestionStatistics(statsCalculation, question);
+    if (statsCalculation.responses.length === 0) {
+      // skip stats for no response
+      return [];
+    }
+    statsCalculation.calculateStatistics();
+
+    statsRows.push(['Option', 'Ranks Received', 'Overall Rank']);
+
+    Object.keys(statsCalculation.ranksReceivedPerOption).sort().forEach((option: string) => {
+      statsRows.push([
+        option,
+        statsCalculation.ranksReceivedPerOption[option].join(', '),
+        statsCalculation.rankPerOption[option] ? String(statsCalculation.rankPerOption[option]) : '',
+      ]);
+    });
+
+    return statsRows;
   }
 
 }
